Refactoring Fuzzing methods in seperate class
